@@ -12,34 +12,34 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.1.3",
-    "@angular/common": "^8.1.3",
-    "@angular/compiler": "^8.1.3",
-    "@angular/core": "^8.1.3",
-    "@angular/forms": "^8.1.3",
-    "@angular/platform-browser": "^8.1.3",
-    "@angular/platform-browser-dynamic": "^8.1.3",
-    "@angular/router": "^8.1.3",
+    "@angular/animations": "^9.0.7",
+    "@angular/common": "^9.0.7",
+    "@angular/compiler": "^9.0.7",
+    "@angular/core": "^9.0.7",
+    "@angular/forms": "^9.0.7",
+    "@angular/platform-browser": "^9.0.7",
+    "@angular/platform-browser-dynamic": "^9.0.7",
+    "@angular/router": "^9.0.7",
     "clone": "^2.1.2",
     "core-js": "^2.5.7",
     "immutable": "^4.0.0-rc.12",
     "memo-decorator": "^1.0.2",
     "ngx-auto-unsubscribe": "^2.4.1",
-    "rxjs": "^6.5.2",
+    "rxjs": "^6.5.4",
     "subsink": "^1.0.0",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.9.1"
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.800.0",
-    "@angular/cli": "^8.0.2",
-    "@angular/compiler-cli": "^8.1.3",
-    "@angular/language-service": "^8.1.3",
+    "@angular-devkit/build-angular": "~0.900.7",
+    "@angular/cli": "^9.0.7",
+    "@angular/compiler-cli": "^9.0.7",
+    "@angular/language-service": "^9.0.7",
     "@types/clone": "^0.1.30",
     "@types/jasmine": "~2.8.7",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~10.12.0",
-    "codelyzer": "^5.0.1",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "jasmine-core": "~3.2.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "^4.2.0",
@@ -51,6 +51,6 @@
     "protractor": "~5.4.1",
     "ts-node": "~7.0.1",
     "tslint": "~5.11.0",
-    "typescript": "3.4.5"
+    "typescript": "3.7.5"
   }
 }
