@@ -1,16 +1,16 @@
 import { NgModule, Optional, SkipSelf } from '@angular/core';
 import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
-
-import { DataService } from './services/data.service';
-import { ClonerService } from './services/cloner.service';
-import { EventBusService } from './services/event-bus.service';
-import { HttpClientRxJSService } from './services/httpClientRxJS.service';
 import { AuthInterceptor } from './interceptors/auth.interceptor';
+// import { DataService } from './services/data.service';
+// import { ClonerService } from './services/cloner.service';
+// import { EventBusService } from './services/event-bus.service';
+// import { HttpClientRxJSService } from './services/httpClientRxJS.service';
+
 
 @NgModule({
   imports: [ HttpClientModule ],
   declarations: [],
-  providers: [ DataService, ClonerService, EventBusService, HttpClientRxJSService,
+  providers: [ /* DataService, ClonerService, EventBusService, HttpClientRxJSService, */
     { provide: HTTP_INTERCEPTORS, useClass: AuthInterceptor, multi: true  }
   ]
 })
