@@ -2,7 +2,9 @@ import { Injectable } from '@angular/core';
 import { Subject, Subscription, Observable } from 'rxjs';
 import { filter, map } from 'rxjs/operators';
 
-@Injectable()
+@Injectable({
+  providedIn: 'root'
+})
 export class EventBusService {
 
     private subject$ = new Subject();
